Rename endpoint and shift codes to utils
