+++ api/index.py
@@ -6,14 +6,7 @@ from api.models import *
 from api.utils.input_serializer import *
 from api.utils.output_serializer import *
 from api.utils.json_response_to_table import *
-from flask import send_file
-
-
-from json import dumps
-import plotly.graph_objects as go
-
-
-from io import BytesIO
+from api.utils.create_figure import create_figure
 
 
 @app.route("/")
@@ -73,12 +66,8 @@ def get_json_response():
             return jsonify(json_response)
 
 
-@app.route("/get_plot")
+@app.route("/graph")
 def get_image():
-    # app = QtWidgets.QApplication()
-    # win = QtWidgets.QWidget()
-    # win.show()
-    # app.exec()
     if request.method == "GET":
         cities, countries = region_input_manager(json.loads(request.args.get("Region")))
         years = year_input_manager(json.loads(request.args.get("Year")))
@@ -86,57 +75,15 @@ def get_image():
             Population.year.in_(years), Population.country.in_(countries)
         )
         json_response = serialize_queryset(queryset)
+        # Create two dictionaries to store the array corresponding to coutries. Will look like:
+        # country_year["India"] = [2010, 2011, ...]
         country_year = {}
-        country_pop = {}
-        years_array = []
-        pop_array = []
+        country_population = {}
         for obj in json_response:
             if obj["country"] not in country_year:
                 country_year[obj["country"]] = []
-            if obj["country"] not in country_pop:
-                country_pop[obj["country"]] = []
+            if obj["country"] not in country_population:
+                country_population[obj["country"]] = []
             country_year[obj["country"]].append(int(obj["year"]))
-            country_pop[obj["country"]].append(obj["population"])
-
-            years_array.append(obj["year"])
-            pop_array.append(obj["population"])
-
-        # fig = create_figure(years_array,pop_array)
-        return create_fig(country_year, country_pop)
-
-
-def create_figure(country_year, country_pop):
-    fig = Figure()
-
-    axis = fig.add_subplot(1, 1, 1)
-
-    for country in country_year:
-        xpoints = np.array(country_year[country])
-        ypoints = np.array(country_pop[country])
-        axis.plot(xpoints, ypoints)
-    countries = []
-    for country in country_year:
-        countries.append(country)
-    axis.legend(countries)
-
-    return fig
-
-
-def create_fig(country_year, country_pop):
-    fig = go.Figure()
-    layout = go.Layout(
-        title="Population vs Year Index",
-        xaxis=dict(title="Year"),
-        yaxis=dict(title="Population"),
-    )
-    for country in country_year:
-        fig.add_trace(
-            go.Scatter(
-                x=country_year[country],
-                y=country_pop[country],
-                mode="lines",
-                name=country,
-            )
-        )
-    fig.update_layout(layout)
-    return fig.to_html(full_html=False)
+            country_population[obj["country"]].append(obj["population"])
+        return create_figure(country_year, country_population)
